Commit previously modified files

+++ performance_index.h
@@ -1,16 +1,17 @@
 #ifndef PERFORMANCE_INDEX_H
 #define PERFORMANCE_INDEX_H
 #include <string>
+#include <iostream>
 
 struct country
 {
   std::string name;
-  int population;
-  int infections;
-  int deaths;
-  int recovered;
-  int economy;
-  int pi;
+  double population;
+  double infections;
+  double deaths;
+  double recovered;
+  double economy;
+  double pi;
 };
 
 void calculate_performance_by_country(struct country AllCountries[]);
